refactor: update users structure
