+++ src/users/user.service.ts
@@ -4,16 +4,17 @@ import {
   HttpStatus,
   HttpException,
 } from '@nestjs/common';
-import { CreateUserDto, UpdatePasswordDto, User, User_Created } from '../types';
+import { User } from '../types';
+import { CreateUserDto, UpdatePasswordDto, User_Created } from './dto/user.dto';
 import { Data } from 'src/data/data.service';
 import { v4 as uuidv4 } from 'uuid';
 
 @Injectable()
 export class UserService {
-  constructor(private readonly database: Data) {}
+  constructor(private readonly data: Data) {}
 
   async findAll(): Promise<User[]> {
-    const users = await this.database.users;
+    const users = this.data.users;
     if (!users) {
       throw new NotFoundException('Users are not found');
     }
@@ -21,7 +22,7 @@ export class UserService {
   }
 
   async findOne(id: string): Promise<User_Created> {
-    const user = await this.database.users.find((item) => item.id === id);
+    const user = this.data.users.find((item) => item.id === id);
     if (!user) {
       throw new NotFoundException('User is not found');
     }
@@ -36,7 +37,7 @@ export class UserService {
       createdAt: Date.now(),
       updatedAt: Date.now(),
     };
-    this.database.users.push(newUser);
+    this.data.users.push(newUser);
     const result = new User_Created(newUser);
     delete result.password;
     return result;
@@ -46,25 +47,25 @@ export class UserService {
     id: string,
     updatePasswordDto: UpdatePasswordDto,
   ): Promise<User_Created> {
-    const i = this.database.users.findIndex((item) => item.id === id);
+    const i = this.data.users.findIndex((item) => item.id === id);
     if (i === -1) {
       throw new HttpException('User is not found', HttpStatus.NOT_FOUND);
     }
-    if (this.database.users[i].password !== updatePasswordDto.oldPassword) {
+    if (this.data.users[i].password !== updatePasswordDto.oldPassword) {
       throw new HttpException('Old password is wrong', HttpStatus.FORBIDDEN);
     }
-    this.database.users[i].password = updatePasswordDto.newPassword;
-    this.database.users[i].version += 1;
-    this.database.users[i].updatedAt = Date.now();
-    const result = new User_Created(this.database.users[i]);
+    this.data.users[i].password = updatePasswordDto.newPassword;
+    this.data.users[i].version += 1;
+    this.data.users[i].updatedAt = Date.now();
+    const result = new User_Created(this.data.users[i]);
     delete result.password;
     return result;
   }
 
   remove(id: string): boolean {
-    const i = this.database.users.findIndex((item) => item.id === id);
+    const i = this.data.users.findIndex((item) => item.id === id);
     if (i !== -1) {
-      this.database.users.splice(i, 1);
+      this.data.users.splice(i, 1);
       return true;
     }
     throw new NotFoundException('User is not found');
